Initial implementation of Scala-style implicit functions.
Following the style of the `fancy-app` package. Does not handle
operators yet, nor does it handle `.`.

This makes `f(_,1)` expand to `fun(x): f(x,1)`.

Main design question is the behavior of multiple _ in the application.
This PR causes each _ to become a distinct function argument, in
order of their appearance. Alternative designs:
 - multiple _ all become the same binding
 - multiple _ are illegal